Encode and decode strings as Modified UTF-8.
As it turns out, vanilla Minecraft (which uses Java's serialization)
does not write proper UTF-8 for values outside of the first 0xffff
Unicode codepoints. This commit changes the internal encoding and
decoding to match that of vanilla Minecraft, which is "Modified UTF-8".

Note that we permissively handle true UTF-8 input, but always write
MUTF-8 output.

A simple test case is included.

Closes #39.